update script for building and testing

+++ scripts/clean.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
 cd ..
-rm -r build
+
+if [ -d "build" ]; then
+    echo "Removing build directory..."
+    rm -r build
+    echo "Build directory removed."
+else
+    echo "Build directory does not exist. Skipping removal."
+fi
+
 cd scripts
